Extract a base class for tests that use pantsd.
